fix(det): NaN divergence stabilization — clamp FCOS reg, fp32 loss, GIoU inf guard, grad-clip, NaN-guard
P30-Det(det_P30_v2) diverged to loss=NaN at ep20~22. Root cause: unbounded
FCOS reg output (fcos_head.py) → AMP fp16 overflow → inf → GIoU nan → shared
backbone contaminates cls_logits. Amplified by missing grad-clip / NaN-guard.

- fcos_head: clamp reg_out to fp16-safe max (4096)
- losses(FCOS): compute cls/reg/ctr in fp32 (.float()) + GIoU nan_to_num guard
- query_decoder: fp32 CE/box loss inputs
- train_det: clip_grad_norm_(10.0) before optimizer step + skip non-finite-loss batches

AMP kept on (loss islands run fp32). Resume from epoch19_checkpoint (mAP50 0.1384).

+++ objdet/models/heads/query_decoder.py
@@ -203,8 +203,8 @@ class SetCriterion(nn.Module):
     def forward(self, outputs: Dict[str, torch.Tensor],
                 targets: List[Dict[str, torch.Tensor]]) -> Dict[str, torch.Tensor]:
         indices = self.matcher(outputs, targets)
-        pred_logits = outputs['pred_logits']         # (B,N,C+1)
-        pred_boxes = outputs['pred_boxes']           # (B,N,4)
+        pred_logits = outputs['pred_logits'].float()  # (B,N,C+1) — fp32 for stable CE under AMP
+        pred_boxes = outputs['pred_boxes'].float()    # (B,N,4)
         B, N = pred_logits.shape[:2]
         device = pred_logits.device
